Skip deleteing the SSO resources if  is set to true

+++ src/test/java/io/openshift/booster/SecuredBoosterIT.java
@@ -107,10 +107,12 @@ public class SecuredBoosterIT {
   // SSO server cleanup
   @AfterClass
   public static void deleteSSO() {
-    COMMAND_EXECUTOR.execCommand("oc delete all --selector application=sso");
-    COMMAND_EXECUTOR.execCommand("oc delete secret sso-app-secret");
-    COMMAND_EXECUTOR.execCommand("oc delete secret sso-demo-secret");
-    COMMAND_EXECUTOR.execCommand("oc delete serviceaccount sso-service-account");
+    if (!Boolean.getBoolean("skip.sso.init")) {
+      COMMAND_EXECUTOR.execCommand("oc delete all --selector application=sso");
+      COMMAND_EXECUTOR.execCommand("oc delete secret sso-app-secret");
+      COMMAND_EXECUTOR.execCommand("oc delete secret sso-demo-secret");
+      COMMAND_EXECUTOR.execCommand("oc delete serviceaccount sso-service-account");
+    }
   }
 
   private String getToken(String username, String password) {
